pyproject.toml
[project]
name = "omni-comment"
description = "Combine outputs from many jobs into a single comment"
dynamic = ["version"]
readme = "README.md"
license = "ISC"
authors = [{ name = "Mark Skelton" }]
keywords = ["github", "comment"]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/mskelton/omni-commen-py"
Issues = "https://github.com/mskelton/omni-comment-py/issues"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.version]
source = "vcs"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
